Update type of AgentLookingForGroup.DutyCategory

+++ FFXIVClientStructs/FFXIV/Client/UI/Agent/AgentLookingForGroup.cs
@@ -147,7 +147,7 @@ public unsafe partial struct AgentLookingForGroup {
         [FieldOffset(0x3B0), FixedSizeArray(isString: true)] internal FixedSizeArray192<byte> _comment;
     }
 
-    public enum DutyCategory : ushort {
+    public enum DutyCategory : uint {
         None = 0,
         Roulette = 1 << 1,
         Dungeons = 1 << 2,
